Add tests for recent_meter_run field, organize serializers
