+++ datastore/tests/views/test_project.py
@@ -1,6 +1,11 @@
 from .shared import OAuthTestCase
-
+from django.contrib.auth.models import User
+from datastore import models
+from datetime import datetime, timedelta
+import pytz
 import json
+import numpy as np
+from numpy.testing import assert_allclose
 
 class ProjectAPITestCase(OAuthTestCase):
 
@@ -192,3 +197,113 @@ class ProjectAPITestCase(OAuthTestCase):
         response.data[0]['message'] == "could not convert string to float: 'NOT A NUMBER'"
         response.data[0]['project_id'] == 'PROJECT_ID'
         response.data[0]['status'] == 'error - bad field value - update'
+
+    def _create_project_fixture(self):
+        self.user = User.objects.create_user('john', 'lennon@thebeatles.com', 'johnpassword')
+
+        self.empty_project = models.Project.objects.create(
+            project_owner=self.user.projectowner,
+            project_id="PROJECTID_1",
+        )
+
+        self.complete_project = models.Project.objects.create(
+            project_owner=self.user.projectowner,
+            project_id="PROJECTID_2",
+            baseline_period_start=None,
+            baseline_period_end=datetime(2012, 1, 1, tzinfo=pytz.UTC),
+            reporting_period_start=datetime(2012, 1, 2, tzinfo=pytz.UTC),
+            reporting_period_end=None,
+            zipcode="91104",
+            weather_station="722880",
+            latitude=0,
+            longitude=0,
+        )
+
+        self.cm_ng = models.ConsumptionMetadata.objects.create(
+            project=self.complete_project,
+            fuel_type="NG",
+            energy_unit="THM",
+        )
+
+        self.cm_e = models.ConsumptionMetadata.objects.create(
+            project=self.complete_project,
+            fuel_type="E",
+            energy_unit="KWH",
+        )
+
+        records = []
+
+        for i in range(0, 700, 30):
+            if i % 120 == 0:
+                value = np.nan
+            else:
+                value = 1.0
+            records.append(models.ConsumptionRecord(
+                metadata=self.cm_ng,
+                start=datetime(2011, 1, 1, tzinfo=pytz.UTC) + timedelta(days=i),
+                value=value,
+                estimated=False,
+            ))
+
+        for i in range(0, 6000):
+            if i % 4 == 0:
+                value = np.nan
+            else:
+                value = 1
+            records.append(models.ConsumptionRecord(
+                metadata=self.cm_e,
+                start=datetime(2011, 12, 1, tzinfo=pytz.UTC) + timedelta(seconds=i*900),
+                value=value,
+                estimated=False,
+            ))
+
+        models.ConsumptionRecord.objects.bulk_create(records)
+
+    def test_project_with_meter_runs(self):
+        self._create_project_fixture()
+
+        complete_project_data = self.get(
+            '/api/v1/projects/{}/?with_meter_runs=True'
+            .format(self.complete_project.pk)
+        ).json()
+
+        assert complete_project_data['project_id'] == 'PROJECTID_2'
+        assert complete_project_data['recent_meter_runs'] == []
+
+        self.complete_project.run_meter()
+
+        complete_project_data = self.get(
+            '/api/v1/projects/{}/?with_meter_runs=True'
+            .format(self.complete_project.pk)
+        ).json()
+
+        assert complete_project_data['project_id'] == 'PROJECTID_2'
+        rmr = complete_project_data['recent_meter_runs']
+        assert len(rmr) == 2
+        assert_allclose(rmr[0]['annual_savings'], 0, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[0]['gross_savings'], 0, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[0]['annual_usage_baseline'], 12.1666, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[0]['annual_usage_reporting'], 12.1666, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[0]['cvrmse_baseline'], 0, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[0]['cvrmse_reporting'], 0, rtol=1e-3, atol=1e-3)
+        assert rmr[0]['consumption_metadata'] == self.cm_ng.pk
+        assert rmr[0]['project'] == self.complete_project.pk
+
+        assert_allclose(rmr[1]['annual_savings'], 0, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[1]['gross_savings'], 0, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[1]['annual_usage_baseline'], 26280, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[1]['annual_usage_reporting'], 26280, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[1]['cvrmse_baseline'], 0, rtol=1e-3, atol=1e-3)
+        assert_allclose(rmr[1]['cvrmse_reporting'], 0, rtol=1e-3, atol=1e-3)
+        assert rmr[1]['consumption_metadata'] == self.cm_e.pk
+        assert rmr[1]['project'] == self.complete_project.pk
+
+        all_project_data = self.get(
+            '/api/v1/projects/?with_meter_runs=True'
+        ).json()
+
+        assert len(all_project_data) == 3
+        assert len(all_project_data[0]["recent_meter_runs"]) == 0
+        assert len(all_project_data[1]["recent_meter_runs"]) == 0
+        assert len(all_project_data[2]["recent_meter_runs"]) == 2
+
